+++ Sources/EpoxyLayoutGroups/Providers/ReflowsForAccessibilityTypeSizeProviding.swift
@@ -35,6 +35,6 @@ extension EpoxyModeled where Self: ReflowsForAccessibilityTypeSizeProviding {
   // MARK: Private
 
   private var reflowsForAccessibilityTypeSizeProperty: EpoxyModelProperty<Bool> {
-    .init(keyPath: \Self.reflowsForAccessibilityTypeSizes, defaultValue: true, updateStrategy: .replace)
+    .init(keyPath: \ReflowsForAccessibilityTypeSizeProviding.reflowsForAccessibilityTypeSizes, defaultValue: true, updateStrategy: .replace)
   }
 }